# src/Indice.Features.Agents.App/src/app/core/services/dex-openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Dex REST API",
    "contact": {
      "name": "Indice Developer",
      "url": "https://www.indice.gr"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:2001/"
    }
  ],
  "paths": {
    "/api/ingest/faq": {
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "POST /api/ingest/faq — multipart upload of a single FAQ-format Markdown file.",
        "description": "Accepts multipart/form-data with a `file` part containing FAQ-shaped Markdown (`# Category` separates segments, `## Question` introduces each question, the body until the next `##`/`#`/EOF is the answer). Each Q&A pair becomes one retrieval chunk; HeadingPath surfaces as 'Category > Question'. The file's first `#` is used as the document category — the form-data `category` is used only when the file has none.",
        "operationId": "UploadFaq",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "file": {
                        "$ref": "#/components/schemas/FileParam"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "category": {
                        "type": [
                          "null",
                          "string"
                        ]
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "language": {
                        "type": [
                          "null",
                          "string"
                        ]
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "oauth2": [
              "dex",
              "dex:ingest"
            ]
          },
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileParam": {
        "type": "string",
        "format": "binary"
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "IngestionReport": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Identifier of the persisted document; `null` when the upload was skipped.",
            "format": "uuid"
          },
          "source": {
            "type": "string",
            "description": "The uploaded file name (echoed for client convenience)."
          },
          "chunksCreated": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Number of chunks inserted for this document; 0 when skipped."
          },
          "skipped": {
            "type": "boolean",
            "description": "True when the upload matched an existing document hash and no work was performed."
          },
          "skippedReason": {
            "type": [
              "null",
              "string"
            ],
            "description": "Reason for skipping; populated only when bool IngestionReport.Skipped is true (e.g. `\"unchanged\"`)."
          },
          "replaced": {
            "type": "boolean",
            "description": "True when an existing document with the same source was deleted before inserting the new one."
          }
        },
        "additionalProperties": false,
        "description": "The endpoint response shape for a single-file ingest."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ]
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Identity Server OAuth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://my.indice.gr/connect/authorize",
            "tokenUrl": "https://my.indice.gr/connect/token",
            "refreshUrl": "https://my.indice.gr/connect/token",
            "scopes": {
              "dex": "Access to Dex REST API",
              "dex:chat": "Provides access to Chat API",
              "dex:ingest": "Provides access to the ingestion API"
            }
          }
        }
      }
    }
  },
...
}